When upsampling logarithmic binnings, numerical errors can occur in single precision. If that happens, we "snap" edges back to their original value.