@jllanfranchi 2020-05-03
I realized there's a more straightforward, less error-prone, and faster way to do this: Simply keep the original bin edges when building up the new edges. I tested a version of this in my own branch, I'll create a PR so you can see what I did.

Also, I removed the `'units'` from `_hash_attrs` again. The alternative to that is to just apply the units to the bin edges in the oversampling function before returning them. By adding units to the things listed in _hash_attrs, that means that every time a new binning is created, units must be passed around. This slowed down iterating over bin tuples (~0.03 to 0.04 seconds per bin) and iterating over bins (~0.06 to 0.08 s per bin). Kind of a minor point right now, and this is hardly optimized for speed, but there you go.