@atrettin 2020-05-03
superseded by PR #5